+++ src/app/shared/RelayCard.svelte
@@ -1,16 +1,15 @@
 <script lang="ts">
   import cx from "classnames"
-  import {between, poll} from "hurdak"
-  import {onMount} from "svelte"
   import {fly} from "src/util/transition"
   import {stringToHue, hsl} from "src/util/misc"
   import {modal} from "src/partials/state"
   import Toggle from "src/partials/Toggle.svelte"
   import Rating from "src/partials/Rating.svelte"
   import Anchor from "src/partials/Anchor.svelte"
+  import RelayStatus from "src/app/shared/RelayStatus.svelte"
   import RelayCardActions from "src/app/shared/RelayCardActions.svelte"
 
-  import {user, Nip65, Keys, Network} from "src/app/engine"
+  import {user, Nip65, Keys} from "src/app/engine"
 
   export let relay
   export let rating = null
@@ -19,18 +18,7 @@
   export let hideActions = false
   export let showControls = false
 
-  let statusHover = false
-  let meta = null
-
   const openModal = () => modal.push({type: "relay/detail", url: relay.url})
-
-  onMount(() => {
-    return poll(3000, () => {
-      const socket = Network.pool.get(relay.url, {autoConnect: false})
-
-      meta = socket?.meta
-    })
-  })
 </script>
 
 <div
@@ -45,24 +33,7 @@
       <i class={relay.url.startsWith("ws://") ? "fa fa-unlock" : "fa fa-lock"} />
       <Anchor on:click={openModal}>{Nip65.displayRelay(relay)}</Anchor>
       {#if showStatus}
-        <span
-          on:mouseout={() => {
-            statusHover = false
-          }}
-          on:mouseover={() => {
-            statusHover = true
-          }}
-          class="h-2 w-2 cursor-pointer rounded-full bg-gray-6"
-          class:bg-gray-6={!meta}
-          class:bg-danger={meta && meta.quality <= 0.3}
-          class:bg-warning={meta && between(0.3, 0.7, meta.quality)}
-          class:bg-success={meta && meta.quality > 0.7} />
-        <p
-          class="hidden text-sm text-gray-1 transition-all sm:block"
-          class:opacity-0={!statusHover}
-          class:opacity-1={statusHover}>
-          {meta ? meta.description : "Not connected"}
-        </p>
+        <RelayStatus {relay} />
       {/if}
       {#if rating}
         <div class="px-4 text-sm" in:fly={{y: 20}}>